Fix migration script for v1.2 upgrade (#651)
### Feature or Bugfix
- Bugfix

### Detail
- migration script for upgrade to V1.2 had a mistake and is affecting
one customer. Basically the `devStrategy` and `devStages` values were
not backfilled which causes nulls in the RDS table that are not allowed
as this column should contain only non-null values.

In this PR we modify that script for customers that have not updated
yet. It is not 100% clear to me whether we should merge it, but I wanted
to raise awareness of this issue here.

### Relates
- #637 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.