All modules can be disabled at a time (#643)
There was a problem if we wanted to disable all modules at once:
A GraphQL union type cannot be empty. We resolve all unions
programmatically depending whether a module is present or not. Some
unions consist of module related parts only and, when all modules are
disabled, the application fails to start.
To fix the issue we could try not to create a unions only, but the rest
schema depends on it. To solve the issue, the following modules have
been introduced: `feeds`, `catalog` and `vote`. Those are not standalone
modules, but just dependencies, hence if no module using them is present
they won't be loaded

Also, removed the feeds from worksheets