Fix module dependency loading. (#675)
The corresponding module interface for a dependency should be loaded
locally

- Bugfix

There is a issue with the loading of modules for cdkproxy task. The
issue appears since the loading of dependency happens in the module
scope, but should be local for each `ModuleInterface`

For instance, when there is `from dataall.modules.catalog import api` in
`dataall.modules.datasets.__init__` the Catalog API will be loaded for
each dataset module interfaces.



By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.