Fix Get Dataset Table Overview for Non-Dataset Owners (#858)
### Feature or Bugfix
<!-- please choose -->
- Bugfix


### Detail
- If a User who is not Dataset Owner, Env Admin, or Dataset Steward
tries to view a dataset table from the Catalog they cannot view the any
of the tabs for: `Preview`, `Overview`, `Columns`, or `Metrics` and
receive an error: `An error occurred (UnauthorizedOperation) when
calling GET_DATASET_TABLE operation`


- This PR:
- Removes `GlueTableProperties` from being returned in the
`getDatasetTable` query as it is not required for any Dataset Table
operations on data.all's UI
    - Resolves Requester User can see Overview Tab
- Adds Check on Dataset Confidentiality Tag to determine to show data in
`Columns` Tab or not and removes checks on
- Resolves Requester User can see Column Tab (for Unclassified Datasets)
- Removes Resource Permissions checks on `Metrics` Tab queries to only
handle check on Dataset Confidentiality Tag
- Resolves Requester User can see Metrics Tab (for Unclassified
Datasets)
  - Fix bug in update column description
    - Resolves Dataset Owner being able to update Column Description

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.