Added github workflows on release branches and fix migration workflow issues (#860)
### Feature or Bugfix
- Feature
- Bugfix

### Detail
- Add all applicable GitHub workflows to PRs pointing at `v2m*` branches
- fix semgrep finding issue from GitHub workflows from migration script
for notifications type --> added `nosemgrep` as no user input is passed
to the SQL query and only code administrators will have access to the
query.
- fix migration validation: this one is tricky as it succeeds when
running it locally and on a real pipeline. It turns out that the issue
was not on the migration script itself but on the way we dropped and
updated tables in the validation migration stage. For dropping tables,
we were using a different schema that the one used in upgrade database.
This PR removes the schema_name variable and uses the envname as schema
for all cases. One final note, this issue might be related to #788.

Here some screenshots of the resulting local schema for the notification
table after running `make drop-tables` and `make upgrade-db`
<img width="962" alt="image"
src="https://github.com/awslabs/aws-dataall/assets/71252798/0d020d7b-915c-436f-a767-8290d0ac3480">


### Relates
- V2.1 release

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.