feat: update diagrams and add region info in deployment pre-reqs + docs for v2.1.0 (#821)
### Feature or Bugfix
- Feature

### Detail
- Add regions info in deploy to AWS guide
- Replace backend diagrams with the correct relationship between AWS
Worker Lambda and ECS tasks. In the original diagram it looks like the
GraphQL API handler Lambda triggers the ECS tasks, but in fact the ECS
tasks are triggered by the AWS Worker Lambdas

### Relates
-  #791 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).
`N/A`


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Noah Paige <69586985+noah-paige@users.noreply.github.com>
Co-authored-by: Daniel Lorch <98748454+lorchda@users.noreply.github.com>
Co-authored-by: Tejas Rajopadhye <71188245+TejasRGitHub@users.noreply.github.com>
Co-authored-by: Noah Paige <noahpaig@amazon.com>